+++ jscore.js
@@ -2074,6 +2074,36 @@ document.addEventListener || new function () {
 
 });
 
+window instanceof Object || new function () {
+
+	var proto = CSSStyleDeclaration.prototype;
+
+	function toUpperCase(str) {
+		return str.charAt(1).toUpperCase();
+	}
+
+	function toCamelCase(propName) {
+		return propName.replace(/-./g, toUpperCase);
+	}
+
+	Object.defineProperty(proto, "cssFloat", {
+		get: function () {
+			return this.styleFloat;
+		}
+	});
+
+	Object.defineProperty(proto, "getPropertyValue", {
+		value: function (propName) {
+			propName = propName.toLowerCase();
+			if ("float" == propName) {
+				return this.styleFloat;
+			}
+			return this[toCamelCase(propName)];
+		}
+	});
+
+};
+
 window.getComputedStyle || new function () {
 
 	//https://github.com/es-shims/es5-shim/issues/152
@@ -2094,7 +2124,10 @@ window.getComputedStyle || new function () {
 
 	function getPropertyValue(propName) {
 		propName = propName.toLowerCase();
-		return this["float" == propName ? "cssFloat" : toCamelCase(propName)];
+		if ("float" == propName) {
+			return this.cssFloat;
+		}
+		return this[toCamelCase(propName)];
 	}
 
 	function createPropDesc(obj, propName) {
@@ -2425,6 +2458,41 @@ lib.css = {
 			return undefined;
 		};
 
+	},
+
+	get: function (element, property, style) {
+		//todo hyphen-style → camelCase
+		var prefix = this.prefix;
+		if (!style) {
+			style = getComputedStyle(element);
+		}
+		if (Array.isArray(property)) {
+			return property.reduce(function (properties, property) {
+				properties[property] = style[prefix(property)];
+				return properties;
+			}, {});
+		}
+		return style[prefix(property)];
+	},
+
+	set: function (element, properties, computedStyle) {
+		//todo hyphen-style → camelCase
+		var animations, style = element.style, prefix = this.prefix;
+		if (!computedStyle) {
+			computedStyle = getComputedStyle(element);
+		}
+		animations = computedStyle[this.animationName];
+		if (Object(properties) === properties) {
+			Object.keys(properties).forEach(function (property) {
+				var value = properties[property];
+				property = prefix(property);
+				if (computedStyle[property] != value) {
+					style[property] = value;
+				}
+			});
+		}
+		//todo if modified
+		return lib.event.awaitTransAnimEnd(element, animations);
 	}
 
 };