+++ changelog.md
@@ -2,7 +2,8 @@
 
 ##v0.4.0
 
- -
+ - new: methods `.get()` and `.set()` for CSS properties
+ - new: `.cssFloat` and `.getPropertyValue()` for `CSSStyleDeclaration` in IE8
 
 ##v0.3.0